requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI/ML libraries
google-genai>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0

# Data processing (used by geo_audit_agent)
pandas>=2.0.0
numpy>=1.24.0

# Optional: Only needed if using real API calls or advanced features
# openai>=1.0.0
# anthropic>=0.7.0
